% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the Distribution}
\usage{
plot_distribution(df)
}
\arguments{
\item{df}{The dataframe}
}
\value{
A list of ggplot objects from the dataframe
}
\description{
This function creates histograms and boxplots for all numeric columns a data frame
}
\examples{
df <- data.frame(a = rnorm(100), b = rnorm(100))
plot_distribution(df)
}
