% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{qview}
\alias{qview}
\title{Quick View of Data}
\usage{
qview(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A list containing summary statistics and distribution plots of the data frame
}
\description{
This function provides a summary and distribution plots for a data frame
}
\examples{
df <- data.frame(a = rnorm(100), b = rnorm(100))
qview(df)
}
