% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qview.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(df)
}
\arguments{
\item{df}{The data frame}
}
\value{
The data frame with summarys for numeric columns
}
\description{
This function gives a basic summary for numeric columns in a data frame
}
\examples{
df <- data.frame(a = rnorm(100), b = rnorm(100))
summarize_data(df)
}
